// openpectus/frontend/src/app/api/models/ProcessValueCommandChoiceValue.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type ProcessValueCommandChoiceValue = {
    value: string;
    value_type: ProcessValueCommandChoiceValue.value_type;
    options: Array<string>;
};

export namespace ProcessValueCommandChoiceValue {

    export enum value_type {
        CHOICE = 'choice',
    }


}